Fix cyclic-import
